Modified.

+++ App/InSituVis.mpi.CameraPath/InSituVis.cpp
@@ -21,9 +21,9 @@ const auto Pos = [] ( const float r )
 {
     const auto tht = kvs::Math::pi / 4.0f;
     const auto phi = kvs::Math::pi / 4.0f;
-    const auto x = r * std::sin( tht ) * std::sin( phi );
-    const auto y = r * std::cos( tht );
-    const auto z = r * std::sin( tht ) * std::cos( phi );
+    const auto x = static_cast<float>( r * std::sin( tht ) * std::sin( phi ) );
+    const auto y = static_cast<float>( r * std::cos( tht ) );
+    const auto z = static_cast<float>( r * std::sin( tht ) * std::cos( phi ) );
     return kvs::Vec3{ x, y, z };
 };
